refactor(redskilled): extract the tolerant lane-row decoder to its own module
The #3651 fix pushed event-lane.ts past the 800-line ratchet; the decoder
is its own domain (generation-tolerant decode) and now lives in
event-lane-decode.ts.
